% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCTailTest.R
\name{GCTailTest}
\alias{GCTailTest}
\title{Granger Causality in Tail Test function for a multivariate time series of tail hits.}
\usage{
GCTailTest(Z, maxp = 1, conf_level = 0.05, method = "LL")
}
\arguments{
\item{Z}{A TxN multivariate time series of 0 and 1 indicating whether a tail event occurred for variable \code{i=1,...,N} at time \code{t=1,...,T}.}

\item{maxp}{Maximum order of the BiDAR(p) model to test Tail Granger Causality for. Defaults to 1.}

\item{conf_level}{Desired confidence level for the likelihood ratio test. Defaults to 0.05.}

\item{method}{Estimation method for the BiDAR(p) models. Defaults to "LL" standing for log-likelihood maximization (no other option in this version).}
}
\value{
A list with elements
\describe{
\item{GC_adjacency}{An adjacency matrix for the resulting Tail Granger Causality graph. A 1 in position \code{[i,j]} indicates a rejected TGC test from variable \code{i} to variable \code{j}.}
\item{p.values}{A matrix of p-values for the tests in the same shape as the \code{GC_adjacency} argument.}
\item{order_mat}{A matrix reporting the order at which the Tail Granger Causality was found between the variables. This uses the AIC criterion for optimal BiDAR(p) selection.}
}
}
\description{
Function to test the presence of Granger Causality in tail relations by means of the likelihood ratio test proposed in Mazzarisi et al. (2020).
}
