% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_BiDARp}
\alias{sample_BiDARp}
\title{sampling function for BiDAR(p) model}
\usage{
sample_BiDARp(Z, p, nu, lambda, gamma, chi)
}
\arguments{
\item{p}{The order of the BiDAR(p) model}

\item{nu}{A vector of length 2 containing the values of the copy probability parameters \nu. Each must be in (0,1).}

\item{gamma}{An array of dimension \code{c(2,2,p-1)} containing the probabilities of copying from the \code{p-1} past values of \code{x}, starting from the last row. Each element \code{gamma[i,j,l]} is the probability that variable \code{i} copies from variable \code{j} at lag \code{l}. Each must be in (0,1) and the sum over the last dimension has to be smaller than 1.}

\item{chi}{A vector of length 2 containing the values of the independent sampling probability \chi. Each must be in (0,1).}

\item{x}{A matrix with \code{p} rows and 2 columns containing the initial configurations.}
}
\value{
A sample from the DAR(p) conditional probability distribution.
}
\description{
This function samples an observation from a BiDAR(p) model starting from a set of initial configurations and given parameters.
}
