% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_darp}
\alias{sample_darp}
\title{sampling function for DAR(p) model}
\usage{
sample_darp(x, p, nu, gamma, chi)
}
\arguments{
\item{x}{A vector of length \code{p} containing the initial configurations.}

\item{p}{The order of the DAR(p) model}

\item{nu}{The value of the copy probability parameter \nu. Must be in (0,1).}

\item{gamma}{A vector of length \code{p-1} containing the probabilities of copying from the \code{p-1} past values of \code{x}, starting from the last element. Each must be in (0,1) and the sum has to be smaller than 1.}

\item{chi}{The value of the independent sampling probability \chi. Must be in (0,1).}
}
\value{
A sample from the DAR(p) conditional probability distribution.
}
\description{
This function samples an observation from a DAR(p) model starting from a set of initial configurations and given parameters.
}
